tools/requirements.txt
# 核心依赖
requests>=2.28.0
PyYAML>=6.0
python-dateutil>=2.8.2

# AI模型API客户端
openai>=1.0.0
anthropic>=0.3.0  # Claude API

# 数据处理
pandas>=1.5.0
numpy>=1.24.0

# HTML解析 (用于网页爬虫)
beautifulsoup4>=4.11.0
lxml>=4.9.0

# HTTP客户端增强
httpx>=0.24.0

# 缓存
diskcache>=5.4.0

# 日志
loguru>=0.6.0

# 配置管理
python-dotenv>=1.0.0

# 命令行工具
click>=8.1.0
rich>=13.0.0  # 美化终端输出

# 数据验证
pydantic>=2.0.0

# 正则表达式增强
regex>=2023.6.3

# XML解析 (arXiv API)
xmltodict>=0.13.0

# 速率限制
ratelimit>=2.2.1

# 进度条
tqdm>=4.65.0